requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
google-generativeai>=0.8.0
pdfplumber==0.11.0
gunicorn==21.2.0
